AI-VİDEO/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-generativeai
elevenlabs
ffmpeg-python
requests
google-auth-oauthlib
google-api-python-client
